refactor: shorten import
