+++ src/cli/commands/version.command.ts
@@ -2,8 +2,11 @@ import { readFileSync } from 'node:fs';
 import { resolve } from 'node:path';
 import chalk from 'chalk';
 import { Command } from './command.interface.js';
-import { logError, logInfo } from '../../shared/helpers/console.js';
-import { getErrorMessage } from '../../shared/helpers/index.js';
+import {
+  logError,
+  logInfo,
+  getErrorMessage,
+} from '../../shared/helpers/index.js';
 
 type PackageJSONConfig = {
   version: string;
