+++ src/shared/helpers/index.ts
@@ -1,3 +1,4 @@
 export * from './common.js';
 export * from './offer.js';
 export * from './file-system.js';
+export * from './console.js';
